Cache packages loaded from files
Solver works on cached packages. Packages are cached either manually,
or by exploring all installation paths. To obtain cflags/libs with the
solver we need to cache the packages loaded from files.

Drawbacks:

(1) pkgconf tool flushed the cache before retrieving nonstatic libs
("reset solver when solving for library groups"). I don't understand
why. I was unable to observe any effect. Because we stash the package
into a cache, we cannot drop the cache when retrieving the libs. Thus
this patch removes flushing the cache added in port to 1.9.4.

(2) The from-file-loaded package poisons the cache. When querying
other packages, the from-file-loaded package could be included in the
results. I'm not sure whether the original isolation was good or bad.
I can imagine someone wants to override a system-provided pkgconfig
file and thus loads the override from a file.

Alternatively, we could cache the package loaded from a file
temporarily just around using the solver. That would narrow the time
window when the package is in the cache.