Cache packages loaded from file only temporarily
This implements an alternative approach outlined in
d3efe46b52b6ae3defb90cd695e835ebf6d13204 commit.

This prevents from having packages loaded from files in a client-wide
cache all the time. The cache is used for resolving cflags/libs.

A downside is that the code is more complex.